+++ test/test_computations/test_cbrng.py
@@ -54,6 +54,18 @@ class TestNormalBM:
         dtype = numpy.float64 if double else numpy.float32
         return normal_bm(self.bijection, dtype, mean=self.mean, std=self.std)
 
+class TestNormalBMComplex:
+    def __init__(self, bijection, mean, std):
+        self.extent = None
+        self.mean = mean
+        self.std = std
+        self.bijection = bijection
+        self.name = 'normal_bm_complex'
+
+    def get_sampler(self, double):
+        dtype = numpy.complex128 if double else numpy.complex64
+        return normal_bm(self.bijection, dtype, mean=self.mean, std=self.std)
+
 class TestGamma:
     def __init__(self, bijection, shape, scale):
         self._shape = shape
@@ -123,6 +135,7 @@ def pytest_generate_tests(metafunc):
         vals = [
             TestUniformFloat(bijection, -5, 7.7),
             TestNormalBM(bijection, -2, 10),
+            TestNormalBMComplex(bijection, -3 + 4j, 7),
             TestGamma(bijection, 3, 10)]
         ids = [test.name for test in vals]
         metafunc.parametrize('test_sampler_float', vals, ids=ids)